rename london mock folder
